Random, pointless changes (mostly trailing whitespace)

+++ src/openblocks/common/block/BlockTank.java
@@ -123,9 +123,9 @@ public class BlockTank extends OpenBlock {
 				itemStack.setTagCompound(nbt);
 			}
 			float f = 0.7F;
-			double d0 = world.rand.nextFloat() * f + (1.0F - f) * 0.5;
-			double d1 = world.rand.nextFloat() * f + (1.0F - f) * 0.5;
-			double d2 = world.rand.nextFloat() * f + (1.0F - f) * 0.5;
+			float d0 = world.rand.nextFloat() * f + (1.0F - f) * 0.5F;
+			float d1 = world.rand.nextFloat() * f + (1.0F - f) * 0.5F;
+			float d2 = world.rand.nextFloat() * f + (1.0F - f) * 0.5F;
 			EntityItem entityitem = new EntityItem(world, x + d0, y + d1, z + d2, itemStack);
 			entityitem.delayBeforeCanPickup = 10;
 			world.spawnEntityInWorld(entityitem);
